News_Portal/backend/requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.23

# Authentication
PyJWT==2.8.0

# HTTP Requests
requests==2.31.0

# Web Scraping
beautifulsoup4==4.12.2
feedparser==6.0.10
lxml==4.9.3

# Machine Learning & NLP
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3

# Sentiment Analysis
textblob==0.17.1
vaderSentiment==3.3.2
nltk==3.8.1

# Utilities
python-dotenv==1.0.0

